Rename from speedtest_cli.py to speedtest.py, but maintain backwards compat with a symlink
